System can search for subject requested by student.
When student a tutor specialisation and lesson, the system checks the database to see if the subject is available or not. If subject is available, then the subjectID can be used to make the bid and if subject is not available, then a new subject needs to be created in the database and the id of the new subject will be used.